improving some type organization
