+++ SwiftLSPClient/Types/Window/MessageActionItem.swift
@@ -1,0 +1,16 @@
+//
+//  MessageActionItem.swift
+//  SwiftLSPClient
+//
+//  Created by Matt Massicotte on 2019-09-02.
+//  Copyright © 2019 Chime Systems. All rights reserved.
+//
+
+import Foundation
+
+public struct MessageActionItem {
+    public let title: String
+}
+
+extension MessageActionItem: Codable {
+}
