+++ SwiftLSPClient/Types/Window/ShowMessageParams.swift
@@ -1,0 +1,11 @@
+//
+//  ShowMessageParams.swift
+//  SwiftLSPClient
+//
+//  Created by Matt Massicotte on 2019-09-02.
+//  Copyright © 2019 Chime Systems. All rights reserved.
+//
+
+import Foundation
+
+public typealias ShowMessageParams = LogMessageParams
